+++ src/time/cds.rs
@@ -82,6 +82,7 @@ pub enum LengthOfDaySegment {
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub enum SubmillisPrecision {
     Absent = 0b00,
     Microseconds = 0b01,